The yut thrower in the yuunori(notes: ko-kr)
This is the yut throwing program from the traditional board game of Korea. A player throws 4 sticks called 'Yut' as a dice. After throwing all yuts, player counts the yuts turned to head side. The point gets big as many as yuts turn to head. If no yut turned to head, the point is 5. And one yut exists with the mark to backward. If all yuts turned to tail side without the one, the player gets 1 point to backward.

I wrote the notes in Korean in my code. I will upload one translated to English soon.